requirements.txt
torch>=2.1
torchvision>=0.16
pillow>=10.0
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.8
tqdm>=4.66
